snapm: move size_fmt() from snapm.report to snapm
Allow size_fmt() to be used in other modules for formatting
human-readable sizes.
